Add a simple diffuse lighting model to the raytracing algorithm
The renderer partially supports directional and point light sources as described in the X3D Lighting component. Any number of these light sources can be defined, but only their intensity, direction (for directional lights) and location (for point lights) are taken into account. If a directional light contains an ambient intensity, the value is added to the scene as a distinct "ambient light" source. Moreover, all light is white.