Add a simple specular lighting model to the raytracing algorithm
The renderer supports material shininess as described in the X3D Shape component. The value is interpreted as an exponent of the cosine function used to compute the specular light intensity. Moreover, directional lights are handled correctly with respect to the specification of the X3D Lighting component, which specifies the direction from the light source, while the algorithm uses the direction to the light source.